04/demo/requirements.txt
jupytext
matplotlib
nltk
pathlib
pyyaml
spacy
scikit-learn
# spaCy English model (pre-built wheel; install with same pip/uv as above)
#https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl


